twitter_bookmark_manager/requirements.txt
flask>=3.0.0
tweepy>=4.14.0
sentence-transformers>=2.2.2
chromadb>=0.4.22
sqlalchemy>=2.0.27
transformers>=4.37.2
imagehash>=4.3.1
llama-cpp-python>=0.2.31
python-dotenv>=1.0.0
cryptography>=41.0.7
pytest>=7.4.4
Jinja2>=3.1.3
Werkzeug>=3.0.1
requests>=2.31.0
numpy>=1.24.3
pandas>=2.1.4
tqdm>=4.66.1
pillow>=10.2.0
torch>=2.1.2